change pipefail to failed with rightmost exit status
