+++ crates/nu-experimental/src/options/pipefail.rs
@@ -16,6 +16,6 @@ impl ExperimentalOptionMarker for PipeFail {
     const DESCRIPTION: &'static str = "\
         Raise an error if an external command fails within a pipeline.
         If an external command fails within a pipeline, $env.LAST_EXIT_CODE is set
-        to the exit code of any command which failed.";
+        to the exit code of rightmost command which failed.";
     const STATUS: Status = Status::OptIn;
 }
