+++ crates/nu-protocol/src/pipeline/pipeline_data.rs
@@ -996,22 +996,6 @@ impl PipelineExecutionData {
     }
 }
 
-/// Check the exit status of each pipeline element.
-///
-/// This is used to implement pipefail.
-#[cfg(feature = "os")]
-pub fn check_exit_status_future(
-    exit_status: Vec<Option<(Arc<Mutex<ExitStatusFuture>>, Span)>>,
-) -> Result<(), ShellError> {
-    let mut result = Ok(());
-    for (future, span) in exit_status.into_iter().rev().flatten() {
-        if let Err(err) = crate::process::check_exit_status_future_ok(future, span) {
-            result = Err(err)
-        }
-    }
-    result
-}
-
 impl From<PipelineData> for PipelineExecutionData {
     #[cfg(feature = "os")]
     fn from(value: PipelineData) -> Self {
